converts siri shortcut to Niles GXR-2 command strings
This php page is used to convert siri commands to GXR-2 command strings.  The input comes from zone and device variables passed in via the URL.  The output is a command string that mimics the output from a GXR-2 control panel.  This page can be run on any device with a web server and interfaces to two separate networks:  a LAN and the GXR-2 expansion port network.  This project is being written for the raspberry pi, but other devices can be used.